+++ CupCarbon-master/src/user/UserList.java
@@ -84,9 +84,9 @@ public class UserList {
 			user.setDataEncrypted(SelectedUser.getBoolean("dataEncrypted"));
 			user.setPreferredLatency(SelectedUser.getDouble("preferredLatency"));
 			user.setPreferredThroughput(SelectedUser.getDouble("preferredThroughput"));
-			user.setPreferredFrequency(SelectedUser.getLong("preferredFrequency"));
-			user.startTime = SelectedUser.getLong("startTime");
-			user.endTime = SelectedUser.getLong("endTime");
+			user.setPreferredFrequency(SelectedUser.getDouble("preferredFrequency"));
+			user.startTime = SelectedUser.getInteger("startTime");
+			user.endTime = SelectedUser.getInteger("endTime");
 
 
 			users.add(user);
